+++ src/tabs/consumer.ts
@@ -1,6 +1,7 @@
 import { Observable } from 'rxjs/Observable';
 import { Consumer } from '../common';
 import { tabs } from './index';
+import CreateOptions = tabs.CreateOptions;
 
 const protectedProvidersWeakMap = new WeakMap<TabsConsumer, tabs.TabsProviderInterface>();
 
@@ -26,6 +27,10 @@ export class TabsConsumer extends Consumer implements tabs.TabsConsumer {
 
   nav() {
     return protectedProvidersWeakMap.get(this)!.nav();
+  }
+
+  create(options: CreateOptions) {
+    return protectedProvidersWeakMap.get(this)!.create(options);
   }
 
   getTabWebContentsState(tabId: string) {